Moves Backstabber to separate file
