+++ axelrod/strategies/oncebitten.py
@@ -111,38 +111,3 @@ class ForgetfulFoolMeOnce(Player):
     def reset(self):
         self.D_count = 0
         self.history = []
-
-
-class BackStabber(Player):
-    """
-    A variation of Fool Me Once, this strategy will
-    wait for the fourth defection and also defect on
-    rounds 198 and 199 unconditionally.
-    """
-
-    name = 'BackStabber'
-    memory_depth = float('inf')  # Long memory
-
-    def __init__(self):
-        """
-        Initialised the player
-        """
-        super(BackStabber, self).__init__()
-        self.D_count = 0
-        self._initial = 'C'
-        self.stochastic = False
-
-    def strategy(self, opponent):
-        if len(opponent.history) > 197:
-            return 'D'
-        if not opponent.history:
-            return self._initial
-        if opponent.history[-1] == 'D':
-            self.D_count += 1
-        if self.D_count > 3:
-            return 'D'
-        return 'C'
-
-    def reset(self):
-        self.D_count = 0
-        self.history = []
